move query default tests to their own file and test
